requirements.txt
argcomplete==2.0.0
boto3==1.26.37
botocore==1.29.37
certifi==2022.12.7
cfn-flip==1.3.0
charset-normalizer==2.1.1
click==8.1.3
durationpy==0.5
et-xmlfile==1.1.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
hjson==3.1.0
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.1
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.2
placebo==0.9.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.45
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.1
troposphere==4.2.0
urllib3==1.26.13
Werkzeug==2.2.2
XlsxWriter==3.0.5
zappa==0.56.1
zipp==3.11.0
python-dotenv==0.21.1
Flask-Migrate==4.0.1
Flask-Cors==3.0.10

